template/requirements/base.txt
celery==5.0.5
channels==3.0.3
Django==3.1.6
django-environ==0.4.5
djangorestframework==3.12.2
psycopg2==2.8.6
redis==3.5.3
setuptools==53.0.0